+++ CHANGES.rst
@@ -5,7 +5,7 @@ Version 2.3.8
 
 Unreleased
 
--   Fix parsing of multipart bodies. 
+-   Fix parsing of multipart bodies.
     Adjust index of last newline in data start. :issue:`2761`