+++ CHANGES.rst
@@ -8,6 +8,8 @@ Unreleased
 -   Use ``flit_core`` instead of ``setuptools`` as build backend.
 -   Fix parsing of multipart bodies.
     Adjust index of last newline in data start. :issue:`2761`
+-   ``_plain_int`` and ``_plain_float`` strip whitespace before type
+    enforcement. :issue:`2734`
 
 
 Version 2.3.6